Move additional constants into shared folder
