+++ tests/shared-test-code/utils/zendesk/createZendeskTicket.ts
@@ -4,7 +4,7 @@ import {
   ZENDESK_BASE_URL,
   ZENDESK_END_USER_EMAIL,
   ZENDESK_REQUESTS_ENDPOINT
-} from '../../../integration-tests/constants/zendeskParameters'
+} from '../../constants/zendeskParameters'
 import { ZendeskRequestData } from '../../../integration-tests/types/zendeskRequestData'
 
 export const createZendeskTicket = async (requestData: ZendeskRequestData) => {
