+++ tests/shared-test-code/utils/zendesk/getZendeskTicket.ts
@@ -4,7 +4,7 @@ import {
   ZENDESK_AGENT_EMAIL,
   ZENDESK_BASE_URL,
   ZENDESK_TICKETS_ENDPOINT
-} from '../../../integration-tests/constants/zendeskParameters'
+} from '../../constants/zendeskParameters'
 
 export const getZendeskTicket = async (ticketId: string) => {
   try {
