+++ tests/shared-test-code/utils/zendesk/zendeskTicketComments.ts
@@ -4,7 +4,7 @@ import {
   ZENDESK_AGENT_EMAIL,
   ZENDESK_BASE_URL,
   ZENDESK_TICKETS_ENDPOINT
-} from '../../../integration-tests/constants/zendeskParameters'
+} from '../../constants/zendeskParameters'
 import { ZendeskComment } from '../../../integration-tests/types/zendeskComment'
 
 export const listZendeskTicketComments = async (
